feat: added text slider to banner
