+++ package.json
@@ -27,6 +27,8 @@
     "react-ga": "^2.6.0",
     "react-helmet": "^5.2.1",
     "react-simple-chatbot": "^0.6.1",
+    "react-slick": "^0.24.0",
+    "slick-carousel": "^1.8.1",
     "styled-components": "^4.3.2"
   },
   "devDependencies": {
